python/requirements.txt
python-snap7~=1.2
PyYAML~=6.0
pyrealsense2~=2.53.1.4623
numpy~=1.24.2
opencv-python~=4.7.0.72
pyshine~=0.0.9
PyGObject~=3.44.1